Renaming, cleanup, document and better explain bug
